done work

+++ services/contactsServices.js
@@ -1,25 +1,22 @@
 import { Contact } from "../models/contactModel.js";
 
-export function listContacts() {
-	// Contact.find().select('+email');
-	// Contact.find().select('-name');
-	// Contact.find().select('name phone');
-	return Contact.find();
+export function listContacts(owner) {
+	return Contact.find({ owner });
 }
 
-export function getContactById(contactId) {
-	return Contact.findById(contactId) || null;
+export function getContactById(contactId, owner) {
+	return Contact.findById(contactId).where("owner").equals(owner) || null;
 }
 
-export function removeContact(contactId) {
-	return Contact.findByIdAndDelete(contactId) || null
+export function removeContact(contactId, owner) {
+	return Contact.findByIdAndDelete(contactId).where("owner").equals(owner) || null
 }
 
-export function addContact(name, email, phone) {
-	return Contact.create({ name, email, phone })
+export function addContact(contact) {
+	return Contact.create(contact)
 }
 
-export async function updateContact(contactId, { name, email, phone }) {
+export async function updateContact(contactId, { name, email, phone }, owner) {
 	const contact = await Contact.findById(contactId);
 
 	if (!contact) return null;
@@ -28,19 +25,19 @@ export async function updateContact(contactId, { name, email, phone }) {
 	if (email) contact.email = email;
 	if (phone) contact.phone = phone;
 
-	await contact.save();
+	await contact.save().where("owner").equals(owner);
 
 	return contact;
 }
 
-export async function changeStatusContact(contactId, favorite) {
+export async function changeStatusContact(contactId, favorite, owner) {
 	if (typeof favorite !== 'boolean') return null;
 
 	const contact = await Contact.findById(contactId);
 
 	contact.favorite = favorite;
 
-	await contact.save();
+	await contact.save().where("owner").equals(owner);
 
 	return contact;
 }
